Add default implementations to all FabricBakedModel methods
